requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
google-generativeai==0.3.2
openai==1.6.1
python-dotenv==1.0.0